Change all models to be lists of token instances
Previously, certain values could be single or lists. For example,
there can be 1 contributor or multiple contributors. This made it so
I had to design a solution that automatically decided whether a
specific MEDFORD file had 1 value or multiple.

It has been decided to instead have every single value automatically
be a list, so we do not need to do this engineering.