Add better block recognition
We previously had issues where 'blocks' of information (i.e.
a specific instance of data, a specific instance of paper...)
were recognized by a line of depth 1. For example, the following
correctly created a Cruise block:

@ Cruise ..
@ Cruise-ShipId 42
@ Cruise-ShipName Hitchhiker

However, this would not be identified as a cruise block:

@ Cruise-ShipId 42
@ Cruise-ShipName Hitchhiker

The method of 'block' detection has been updated to instead watch
for changes in either major token, or a decrease in depth, catching
the following block transitions:

@ Data Trek
@ Cruise-ShipName Star

@ Data Trek
@ Data-Type Character
@ Data Star

There still remains the issue of how to recognize blocks of the
same depth and major token, such as:

@ Cruise-ShipId 42
@ Cruise-ShipId 24

These are two separate instances of cruises, but would be read as
one cruise with two ship ids by the current code.