Rename Cruise->Expedition; Adjust Date Validation
Changed 'Cruise' to 'Expedition' for generality. Non-coral data-
sets will likely want to record the data collection, but a cruise
is not used to collect rock samples from the mountains...

Date validation has been changed to no longer accept arbitrary
strings. This was a similarly arbitrary change, and can be
reverted with no current forseen consequences.