Finalize Change to new parsing method
All old logic for parsing lines in a medford file has been officially
deleted. medford is now run by using the src/medford.py script. Test
files have also been updated to use the official extension, .mfd.

Tests have been added to test the medford parsing software,
automatically regenerating the JSON files in the process.

Please note that BCODMO parsing is OFFICIALLY BROKEN as of this commit,
as the parsing structure assumption is no longer accurate. This broken
commit has been deemed acceptable because:
 1. BCODMO parsing has never fully been functional to begin with.
 2. I need to go over the new structure to ensure it makes sense.

As it stands now, I believe the line:
	Data_Ref This Is Some Data
will create a backend that looks as follows:
{Data: [
	{Ref:	[
		 {'desc': "This Is Some Data"}
		]
	}
       ]
}
I am questioning the depth of lists and dictionaries in use, but will
need to examine the amount of code restructuring that will be necessary
to work around this, and whether working around it is really necessary.