Begin implementing BAG copy functionality Started implementing expected BAG translation functionality that copies relevant files into the bag data/ subdirectory structure.
Also implemented functionality to copy remote files identified using the
-Ref minor tag. We may want to add a flag that says "Don't actually
download this file."

Note, the logic needs to be repurposed to not actually download the file
until it has checked that all filenames are correct. Right now, there is
a potential fail case where it downloads the file, THEN finds out it
can't write it for some reason or another, and then leaves the file
sitting around.

Code also must be adjusted to be DRY.