NON-FUNCTIONAL:: Add BagIt folder structure funcs Add various BagIt functionality to its own file, such as moving local files with failure on running out of disk space. Also creates a fetch.txt file.
Add write functionality for the detail_parser, such that it writes a new
file from a dictionary it parsed from an existing Medford file. This may
or may not persist -- it would depend on the difficulty of converting a
pydantic model back to the original dictionary format. It may be
replaced by a method to directly write a pydantic model to a novel
medford file.