Complete refactor of BAGIT - Now minimally functional
Completely refactored the logic behind the bagit translation module of MEDFORD. It now can accept a MEDFORD file describing local and remote files, copy the local files, note down the remote files, create the manifest, create the fetch, AND modify + copy the source medford file into the bag. Note that it does not yet zip up the final bag, but this is hopefully easy.

It is also currently fragile -- I forsee issues with filenames, with out of disk space errors, and similar edge cases.