Change type logic behind parsing details
It didn't sit right with me how I was passing a 4-tuple around of type (bool, bool, Optional, Optional). I changed the typing to make this either a 'detail_return' type (which contains whether or not the detail is novel and the detail itself) or a str (aka it just parsed a macro) or None. So far, it definitely makes the tests easier, as now dr is always punted to the next FromLine call without question...