Add Pretty-Print errors; Refactor Data to persist line no
Firstly:
Changed the way the medford detail parser worked to store the line at
  which some data was parsed into the dict that it passes to pydantic.

Secondly:
Changed pydantic types to be typevars.
Changed all str, int, etc references to be typevar T, and furthermore,
  changed all T references to be references to Tuple[int, T] to capture
  the line number data was parsed at.

Thirdly:
Added functionality to detailparser to capture ValidationErrors thrown
  by Pydantic and make them more user-friendly:
  a. what line of the original file the error is from
  b. the tokens involved in that line
  c. slightly better description of what the error is

Part 3 is still a work in progress, as there are some assumptions
about the numbers of errors and whatnot that are absolutely horrifying.
There are some errors thrown by pydantic that will cause the whole thing
to crash, but mostly these errors come from incorrect implementation in
medford.

(i.e., keep note that in order to validate something like "corresponding
authors need emails, use a root validator in the contributor class,
not a validator in the entity class that contains contributors.)