Add "No Errors" msg;Remove No-Role Contributor bug
Added a message that the medford parser prints if no errors are found
in the provided MEDFORD file.

Removed a bug where medford will crash when checking that a
Contributor has an email if it is marked as a corresponding author,
but the Contributor field in question has no roles.