A Lot : Mostly BagIt Support
Unfortunately, due to my method of development, a lot of things have
been combined into this single git commit.

An overview of the changes are as follows:

1. the error manager can now (barely) support errors that occur while
details are consumed. This needs to be finished.

2. BagIt compilation support has been (mostly) updated. It now expects
the updated internal structure of details, accounting for the line
numbers. This may later need to be restructured to be something that it
doesn't have to worry about, but that is too much object oriented
refactoring for version 1.0. Perhaps it will be taken care of in v1.1.

2.1 File models have been adjusted. Local files now expect the lines
"Path" and "Destination", each of which contain the filename. Remote
files now expect "Filename" and "URI", where Filename is the desired
name of the file if fetched, and URI is the location the file can be
obtained from. This may later be refacted to be "Destination" rather
than "Filename", so that a user can specify a destination in the BagIt
data structure. Outpath storage has also been slightly modified to be
cleaner.

2.2 Minor validation has been added to ArbitraryFiles, ensuring that it
is either a local or a remote file, and a flag is set specifying which
it is.

Currently BagIt compilation is 90% implemented, there is one final
missing step: The MEDFORD internal representation -> new output file
conversion. This conversion was written before the LineNumber
propogation change, so completely collapses on the new internal
representation and needs to be updated to create a correct MEDFORD file
for storage into the Bag.