Update MEDFORD File re-writing
the medford parser is now capable of writing a MEDFORD file in correct
MEDFORD syntax. Example use case: compilation to a bag, where it has to
not only update the locations of files to be the bag-relative location,
but also actually write the novel MEDFORD file into the bag.

This was achieved by abandoning the json-based approach and instead
taking advantage of the functions provided by the pydantic models.

Note that order between version of the file is NOT consistent because
of this. Expected tokens (fields, in pydantic's perspective) are
written in the order they were defined when defining the models. For
example, even though a medford file will say Journal-Pages and then
Journal-Issue, because in the Journal model Issue was defined first,
it will be written as Journal-Issue and then Journal-Pages.

Unexpected tokens will also be ordered inconsistently. This is because
in order to access them, we have to use the __set_fields__ attribute of
a model. The name not only refers to "what fields have been set" but
also "these are represented as a set", meaning that order is not
consistent. This is expected behavior by the pydantic developers, so
I will not persue it further.