Add Syntax Error Management Capabilities
The medford parser is now able to recognize a limited set of syntactical
errors, such as "undefined macro" or "repeated macro name". It will
automatically stop running before running the Pyndatic parsing process
if it encounters any of these errors and print a basic error message
describing the issue to the user.