Add Macro resolution logic; Macro resolution tests Separated "get content" from "resolve" for Macros. To get the content of a Macro after its macros have been replaced, use the "resolve" function.
This is because Macros keep track as to whether or not they've been
resolved before, to save processing time.

For example, if two macros reference Macro1, the second time Macro1
has resolve() called, it should immediately return its .resolution
attribute.

Added a ludicrous amount of tests to try and make sure macro logic
works.

Removed test_obj_lineprocessor, because it's a holdover from before
I renamed it to linecollector.