Documenting the solution to import issues
It turns out SetupTools was right all along. I was testing
medford.py like a script in VSCode, which caused it to require
a different import style than SetupTools, which _properly_
treated the MEDFORD directory as a module.

I've added a note to the bottom of the readme so no one else
makes this extremely long-lived mistake.