Add get_major_minors() function
Add functionality to poll the MEDFORD parser for what major and minor tokens
it is aware of.

To use this, import MEDFORD and use MEDFORD.get_major_minors()

Note that this is manually updated. On adding new major/minor tokens, the
model python file's DefinedMajorMinor must be updated, and the updated
class' `minors` classmethod should be updated.

If a new module of models is added, `models/__init__.py` must be updated to
import this module and update its `models_get_major_minors` function to
update its output with that module's `DefinedMajorMinor`